# Ensure TDA.__init__ validates freq