--- a/pandas/tests/arrays/test_timedeltas.py
+++ b/pandas/tests/arrays/test_timedeltas.py
@@ -11,7 +11,7 @@ import pandas.util.testing as tm
 class TestTimedeltaArrayConstructor(object):
     def test_freq_validation(self):
         # ensure that the public constructor cannot create an invalid instance
-        arr = np.array([0, 0, 1]) * 3600 * 10**9
+        arr = np.array([0, 0, 1], dtype=np.int64) * 3600 * 10**9
 
         msg = ("Inferred frequency None from passed values does not "
                "conform to passed frequency D")