Users should not be able to construct invalid instances with the public constructor.

De-duplicates some code.