@TomAugspurger 2019-01-08
Setup

```python
import numpy as np
import pandas as pd

data = np.arange(10_000, dtype='i8') * 24 * 3600 * 10**9
%timeit pd.arrays.TimedeltaArray(data)
%timeit pd.arrays.TimedeltaArray(data, freq='D')
%timeit pd.timedelta_range("1H", periods=10_000)

```

master

```
>>> %timeit pd.arrays.TimedeltaArray(data)
2.65 µs ± 61.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

>>> %timeit pd.arrays.TimedeltaArray(data, freq='D')
70.3 µs ± 1.7 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit pd.timedelta_range("1H", periods=10_000)
209 µs ± 3.32 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

PR (with master merged in)

```
>>> %timeit pd.arrays.TimedeltaArray(data)
19.6 µs ± 2.67 µs per loop (mean ± std. dev. of 7 runs, 100000 loops each)

>>> %timeit pd.arrays.TimedeltaArray(data, freq='D')
588 µs ± 7.03 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> %timeit pd.timedelta_range("1H", periods=10_000)
135 µs ± 5.29 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

The one with `freq="D"` passed is (70us -> 588us) is expected, and can maybe be improved.

The firs one, no freq, is surprising. (2.65us -> 20us). I'll see where the time is being spent now.