@TomAugspurger 2019-01-08
~80% of `TDA.__init__` is in `_from_sequence`. For that ~75% is in `sequence_to_td64ns`, and ~10% in `_simple_new`.

Unfortunately, nothing in `sequence_to_td64ns` really stands out... The largest offender is

```
   845                                               # Convert whatever we have into timedelta64[ns] dtype
   846         1         21.0     21.0     31.8      if is_object_dtype(data) or is_string_dtype(data):
   847                                                   # no need to make a copy, need to convert if string-dtyped
   848                                                   data = objects_to_td64ns(data, unit=unit, errors=errors)
   849                                                   copy = False
 ```

It's just the cumulative effect of all the `is_<foo>_dtype` checks that's adding up.