@TomAugspurger 2019-01-08
So my request in 
https://github.com/pandas-dev/pandas/pull/24666#pullrequestreview-190080984 is my only blocker right now.