@jbrockmendel 2019-01-08
> It's just the cumulative effect of all the is_<foo>_dtype checks that's adding up.

These are checks we do a ton across the code-base, so I'm definitely on board for optimizing them.

Two options come to mind:
- use `data.dtype.kind == 'O'`, which %timeit is telling me takes ~75ns vs ~1.25us for a is_<foo>_dtype check
- do a lib.infer_dtype call up-front

The data.dtype.kind checks seem more straightforward to me.

> continue to speed up sequence_to_td64ns (possible a fast path argument, when the input is known to be an ndarary[i8] or m8[ns].

certainly moving the i8 and m8[ns] cases to the top of the checks can short-circuit checking for other dtypes.

> split out the freq inference of sequence_to_td64ns into a separate method.

The freq inference in sequence_to_td64ns is pretty trivial.  Do you mean the freq validation in _from_sequence?