@jbrockmendel 2019-01-08
just pushed with faster dtype checks