@jbrockmendel 2019-01-08
> Hows the perf for the non-user-provided-freq case compare to master now?

`__init__` is ~2.8x slower than master, _simple_new is about 2.5x faster

setup
```
In [3]: arr = np.arange(1000)
In [4]: arr2 = arr.view('timedelta64[ns]')
```

PR
```
In [5]: %timeit pd.core.arrays.TimedeltaArray(arr)
The slowest run took 30.48 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 6.04 µs per loop

In [6]: %timeit pd.core.arrays.TimedeltaArray(arr2)
The slowest run took 14.43 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.47 µs per loop

In [7]: %timeit pd.core.arrays.TimedeltaArray._simple_new(arr)
The slowest run took 63.68 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 880 ns per loop

In [8]: %timeit pd.core.arrays.TimedeltaArray._simple_new(arr2)
The slowest run took 24.56 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 893 ns per loop
```

master
```
In [5]: %timeit pd.core.arrays.TimedeltaArray(arr)
The slowest run took 40.33 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 2.21 µs per loop

In [6]: %timeit pd.core.arrays.TimedeltaArray(arr2)
The slowest run took 11.29 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 1.94 µs per loop

In [7]: %timeit pd.core.arrays.TimedeltaArray._simple_new(arr)
The slowest run took 10.87 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 2.5 µs per loop

In [8]: %timeit pd.core.arrays.TimedeltaArray._simple_new(arr2)
The slowest run took 17.49 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 2.18 µs per loop
```