@TomAugspurger 2019-01-09
Just to be clear, my only blocker for the RC was the validation of the dtype in simple_new, which has been fixed.

My blockers for the final release are

1. `DTA/TDA.__init__`not accepts strings like `TimedeltaArray(np.array(['1H', '2H']))`.
2. The `__init__` methods be simple and don't do unnecessary work. We know that they should only accept `ndarray[m8[ns]` (or i8) or a TDA / Series / Index boxing those, so checking for additional dtypes is unnecessary.

So I'm perfectly fine with filing issues and going away.